Improv/conflict

+++ medihub-api/medihub-infrastructure/src/main/java/org/medihub/persistence/appointment_request/AppointmentRequestAdapter.java
@@ -1,21 +1,18 @@
 package org.medihub.persistence.appointment_request;
 
 import lombok.RequiredArgsConstructor;
-import org.medihub.application.exceptions.NotAvailableException;
+import org.medihub.application.exceptions.ForbiddenException;
 import org.medihub.application.exceptions.NotFoundException;
 import org.medihub.application.ports.outgoing.appointment.SaveAppointmentRequestPort;
 import org.medihub.application.ports.outgoing.appointment_request.DeleteAppointmentRequestPort;
 import org.medihub.application.ports.outgoing.appointment_request.GetAllAppointmentRequestsPort;
 import org.medihub.application.ports.outgoing.appointment_request.GetAppointmentRequestForClinicPort;
 import org.medihub.application.ports.outgoing.appointment_request.LoadAppointmentRequestPort;
-import org.medihub.domain.appointment.Appointment;
 import org.medihub.domain.appointment.AppointmentRequest;
-import org.medihub.domain.clinic.ClinicAdmin;
-import org.springframework.dao.PessimisticLockingFailureException;
 import org.springframework.stereotype.Component;
 
-import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
@@ -40,7 +37,10 @@ public class AppointmentRequestAdapter implements
     }
 
     @Override
-    public void deleteAppointmentRequest(Long id) {
+    public void deleteAppointmentRequest(Long id) throws ForbiddenException {
+        Optional<AppointmentRequestJpaEntity> appointmentRequestJpaEntity = appointmentRequestRepository.findById(id);
+        if(!appointmentRequestJpaEntity.isPresent())
+            throw new ForbiddenException();
         appointmentRequestRepository.deleteById(id);
     }
 
@@ -54,11 +54,10 @@ public class AppointmentRequestAdapter implements
         AppointmentRequestJpaEntity appointmentRequestJpaEntity =
                 appointmentRequestRepository.findById(id).orElseThrow(NotFoundException::new);
         return mapper.mapToDomainEntity(appointmentRequestJpaEntity);
-
     }
 
     @Override
-    public AppointmentRequest loadByIdWithLock(Long id) throws NotFoundException, NotAvailableException {
+    public AppointmentRequest loadByIdWithLock(Long id) throws NotFoundException {
         AppointmentRequestJpaEntity appointmentRequestJpaEntity;
         appointmentRequestJpaEntity =
                 appointmentRequestRepository.findByIdWithLock(id).orElseThrow(NotFoundException::new);
